Matching previous commit on other PR for same issue
I changed the name and subject of this commit from the 404 issue to the email server issue. Instead of deleting the account.js file which was apart of the previous solution to this PR before  I changed the name, I am matching it to the solution for the new account.js in the patch-1 branch which solves the issue.